Hoist AttrValue into style.
